feat(core): move base page to new core package
