+++ src/core/__init__.py
@@ -3,6 +3,7 @@ Core package for Coles scraper.
 """
 
 from .base_extractor import HtmlExtractor
+from .base_page import BasePage
 from .fetcher import ColesPageFetcher
 from .webdriver_utils import init_seleniumwire_webdriver, initialize_driver
 from .utils import validate_url
