+++ src/page/__init__.py
@@ -1,0 +1,2 @@
+from .categories import CategoriesPage
+from .products import ProductsPage
