+++ src/page/categories.py
@@ -4,7 +4,7 @@ Page Object Model for Coles browse page.
 
 from selenium.webdriver.common.by import By
 
-from .base import BasePage
+from src.core.base_page import BasePage
 
 
 class CategoriesPageLocators:
