--- src/poms/__init__.py
@@ -1,3 +1,0 @@
-from .base import BasePage
-from .products import ProductsPage
-from .categories import CategoriesPage
